Restructured ItemEventResponse to not have nested classes
