+++ Fuyu.Backend.BSG/DTO/Responses/ItemEventResponse.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Runtime.Serialization;
+using Fuyu.Backend.BSG.ItemEvents.Models;
 
 namespace Fuyu.Backend.BSG.DTO.Responses
 {
@@ -11,36 +12,5 @@ namespace Fuyu.Backend.BSG.DTO.Responses
 
         [DataMember(Name = "warnings")]
         public InventoryWarning[] InventoryWarnings = [];
-
-        [DataContract]
-        public class InventoryWarning
-        {
-            [DataMember(Name = "index")]
-            public int RequestIndex { get; set; }
-
-            [DataMember(Name = "errmsg")]
-            public string ErrorMessage { get; set; }
-
-            [DataMember(Name = "code")]
-            public string ErrorCode { get; set; }
-
-            [DataMember(Name = "msg")]
-            public string String_0
-            {
-                set
-                {
-                    this.ErrorMessage = value;
-                }
-            }
-        }
-
-        public class ProfileChange
-        {
-            [DataMember(Name = "experience")]
-            public int Experience;
-
-            [DataMember(Name = "recipeUnlocked")]
-            public Dictionary<string, bool> UnlockedRecipes = [];
-        }
     }
 }
